Add playlist dropdown component
