+++ src/components/cards/verticalCard/verticalCard.css
@@ -41,22 +41,3 @@
   color: var(--light-bg-clr);
 }
 
-.playlist-stacklist {
-  top: 3rem;
-  right: 0.1rem;
-  width: fit-content;
-  position: absolute;
-  z-index: 2;
-  background-color: var(--custom-dark-clr);
-  color: white;
-}
-
-.playlist-li-item {
-  padding-left: 8px;
-}
-
-.playlist-dropdown-inp {
-  padding: 0.4rem;
-  border-radius: 3px;
-  border: none;
-}
