+++ src/components/index.js
@@ -4,5 +4,6 @@ import Chips from "./chips/Chips";
 import VerticalCard from "./cards/verticalCard/VerticalCard";
 import HorizontalCard from "./cards/horizontalCard/HorizontalCard";
 import LargeHorizontalCard from "./cards/LargeHorizontalCard/LargeHorizontalCard";
+import PlaylistDropdown from "./dropdown/playlistDropdown/PlaylistDropdown";
 
-export { Navbar, Sidebar, Chips, VerticalCard, HorizontalCard , LargeHorizontalCard};
+export { Navbar, Sidebar, Chips, VerticalCard, HorizontalCard , LargeHorizontalCard, PlaylistDropdown};
